04-prompt-engineering/requirements.txt
# Core dependencies
anthropic>=0.40.0
openai>=1.54.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
datasets>=2.14.0

# Evaluation
scikit-learn>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
pydantic>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
